+++ install/automation.sh
@@ -1,0 +1,5 @@
+if [ -f ~/dotfiles-versions/automation.sh ] ;then
+    echo ":: AUTOMATION: automation.sh found"
+    source ~/dotfiles-versions/automation.sh
+    echo
+fi