+++ install/displaymanager-auto.sh
@@ -1,0 +1,6 @@
+echo -e "${GREEN}"
+figlet "Display Manager"
+echo -e "${NONE}"
+echo ":: AUTOMATION: Keep current setup of Display Manager"
+disman=0
+echo