+++ install/dotfiles-auto.sh
@@ -1,0 +1,6 @@
+echo -e "${GREEN}"
+figlet "Folder"
+echo -e "${NONE}"
+dot_folder=$automation_dotfilesfolder
+echo "AUTOMATION: Installation folder set to ~/$automation_dotfilesfolder"
+echo