+++ install/flatpak.sh
@@ -1,12 +1,13 @@
 # ------------------------------------------------------
 # Flatpak
 # ------------------------------------------------------
+echo -e "${GREEN}"
+figlet "Flatpak"
+echo -e "${NONE}"
+
 if [[ $(_isInstalledPacman "flatpak") == 0 ]]; then
     echo ":: Flatpak already installed"
 else
-    echo -e "${GREEN}"
-    figlet "Flatpak"
-    echo -e "${NONE}"
     echo "Flatpak is a framework for distributing desktop applications across various Linux distributions." 
     echo "Flatpak applications: these are the applications the user installs via the flatpak command"
     echo