+++ install/hook.sh
@@ -1,5 +1,5 @@
 # ------------------------------------------------------
-# Copy dotfiles
+# Execute hook.sh
 # ------------------------------------------------------
 if [ -f ~/dotfiles-versions/hook.sh ]; then
 echo -e "${GREEN}"