+++ install/init-pywal.sh
@@ -1,7 +1,9 @@
 # ------------------------------------------------------
 # init pywal with default wallpaper
 # ------------------------------------------------------
-
+echo -e "${GREEN}"
+figlet "Pywal"
+echo -e "${NONE}"
 if [ ! -f ~/.cache/wal/colors-hyprland.conf ]; then
     wal -i ~/wallpaper/default.jpg
     echo "Pywal and templates activated."