+++ install/installer-auto.sh
@@ -1,0 +1,6 @@
+echo -e "${GREEN}"
+figlet "Packages"
+echo -e "${NONE}"
+echo ":: AUTOMATION: New packages will be installed"
+force_install=0
+echo