+++ install/keyboard-auto.sh
@@ -1,0 +1,5 @@
+echo -e "${GREEN}"
+figlet "Keyboard"
+echo -e "${NONE}"
+echo ":: AUTOMATION: Proceed with existing keyboard configuration."
+echo