+++ install/neovim.sh
@@ -1,7 +1,7 @@
 # ------------------------------------------------------
 # Neovim
 # ------------------------------------------------------
-if [ -d ~/dotfiles-versions/$version/.config/nvim ] ;then
+if [ -d ~/dotfiles-versions/$version/dotfiles/.config/nvim ] ;then
     if [ -d ~/$dot_folder/nvim ]; then
         echo -e "${GREEN}"
         figlet "Neovim"