+++ install/post.sh
@@ -1,5 +1,5 @@
 # ------------------------------------------------------
-# Copy dotfiles
+# Execute post.sh
 # ------------------------------------------------------
 if [ -f ~/dotfiles-versions/post.sh ]; then
 echo -e "${GREEN}"