+++ install/yay.sh
@@ -1,12 +1,12 @@
 # ------------------------------------------------------
 # Check if yay is installed
 # ------------------------------------------------------
+echo -e "${GREEN}"
+figlet "yay"
+echo -e "${NONE}"
 if sudo pacman -Qs yay > /dev/null ; then
     echo ":: yay is already installed!"
 else
-    echo -e "${GREEN}"
-    figlet "yay"
-    echo -e "${NONE}"
     echo ":: yay is not installed. Starting the installation!"
     _installPackagesPacman "base-devel"
     SCRIPT=$(realpath "$0")