+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Version 2.9.5RL
 --------------------------------------------------------
 - The folder name and location of the dotfiles folder can be now individual defined during the update and installation.
 - You can copy the dotfiles folder to another location and rename it. The folder can the be activated with the ML4W Welcome App.
+- The installation and update steps can be fully automated with the new automation.sh script.
 - Backlight module to waybar for laptop users added.
 
 Version 2.9.4